# wwwroot/swagger/v1/swagger.json
{
...
  "openapi": "3.0.0",
  "info": {
    "title": "Api Comercial",
    "description": "API para la gestión de documentos comerciales",
    "contact": {
      "name": "Soporte API",
      "email": ""
    },
    "version": "v1"
  },
  "paths": {
...
    "/api/Marcas/marcas": {
      "get": {
        "tags": [
          "Marcas"
        ],
        "operationId": "Marcas_GetMarcas",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MarcaAuto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Marcas"
        ],
        "operationId": "Marcas_InsertMarca",
        "requestBody": {
          "x-name": "marca",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MarcaAuto"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MarcaAuto"
                }
              }
            }
          }
        }
      }
    },
    "/api/Marcas/marcas/{id}": {
      "get": {
        "tags": [
          "Marcas"
        ],
        "operationId": "Marcas_GetMarcaPorId",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MarcaAuto"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Marcas"
        ],
        "operationId": "Marcas_UpdateMarca",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 1
          }
        ],
        "requestBody": {
          "x-name": "marca",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MarcaAuto"
              }
            }
          },
          "required": true,
          "x-position": 2
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Marcas"
        ],
        "operationId": "Marcas_DeleteMarca",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      }
    },
    "/api/Modelos/modelos": {
      "get": {
        "tags": [
          "Modelos"
        ],
        "operationId": "Modelos_GetModelos",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ModeloAuto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Modelos"
        ],
        "operationId": "Modelos_InsertModelo",
        "requestBody": {
          "x-name": "modelo",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModeloAuto"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModeloAuto"
                }
              }
            }
          }
        }
      }
    },
    "/api/Modelos/modelos/{id}": {
      "get": {
        "tags": [
          "Modelos"
        ],
        "operationId": "Modelos_GetModeloPorId",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModeloAuto"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Modelos"
        ],
        "operationId": "Modelos_UpdateModelo",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 1
          }
        ],
        "requestBody": {
          "x-name": "modelo",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModeloAuto"
              }
            }
          },
          "required": true,
          "x-position": 2
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Modelos"
        ],
        "operationId": "Modelos_DeleteModelo",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "MarcaAuto": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "idMarca": {
            "type": "integer",
            "description": "Identificador de la marca de autos (PK)",
            "format": "int32",
            "nullable": true
          },
          "descripcionMarca": {
            "type": "string",
            "description": "Descripción de la marca de autos",
            "nullable": true
          }
        }
      },
      "ModeloAuto": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "idModelo": {
            "type": "integer",
            "description": "Identificador del modelo de auto (PK)",
            "format": "int32",
            "nullable": true
          },
          "idMarca": {
            "type": "integer",
            "description": "Identificador de la marca de auto (FK)",
            "format": "int32"
          },
          "nombreMarca": {
            "type": "string",
            "description": "Nombre de la marca"
          },
          "descripcionModelo": {
            "type": "string",
            "description": "Descripción del modelo de auto"
          }
        }
      },
...
    },
    "securitySchemes": {
      "JWT": {
        "type": "apiKey",
        "description": "Ingrese: Bearer {su token JWT}",
        "name": "Authorization",
        "in": "header"
      }
    }
  },
  "security": [
    {
      "JWT": []
    }
  ]
}